feat: Add LifecycleEntity as abstract base class for managing lifecycle of resources
- Introduced `LifecycleEntity` as an abstract base class to manage the lifecycle of resources.
- Updated `VaradhiTopic` to inherit from `LifecycleEntity` for generic lifecycle management.
- Added lifecycle status handling methods to `LifecycleEntity` for consistent state management across resources.
