+++ entities/src/main/java/com/flipkart/varadhi/entities/VaradhiTopic.java
@@ -12,12 +12,11 @@ import java.util.Map;
  */
 @Getter
 @EqualsAndHashCode(callSuper = true)
-public class VaradhiTopic extends AbstractTopic {
+public class VaradhiTopic extends LifecycleEntity {
 
     private final Map<String, InternalCompositeTopic> internalTopics;
     private final boolean grouped;
     private final TopicCapacityPolicy capacity;
-    private final LifecycleStatus status;
 
     /**
      * Constructs a new VaradhiTopic instance.
@@ -102,81 +101,5 @@ public class VaradhiTopic extends AbstractTopic {
      */
     public InternalCompositeTopic getProduceTopicForRegion(String region) {
         return internalTopics.get(region);
-    }
-
-    /**
-     * Marks the topic as created.
-     */
-    public void markCreated() {
-        this.status.update(LifecycleStatus.State.CREATED);
-    }
-
-    /**
-     * Marks the topic as created with the specified actor code and message.
-     *
-     * @param actorCode the actor code indicating the reason for the state
-     * @param message   the message associated with the state
-     */
-    public void markCreated(LifecycleStatus.ActorCode actorCode, String message) {
-        this.status.update(LifecycleStatus.State.CREATED, message, actorCode);
-    }
-
-    /**
-     * Marks the topic creation as failed with the specified message.
-     *
-     * @param message the message associated with the failure
-     */
-    public void markCreateFailed(String message) {
-        this.status.update(LifecycleStatus.State.CREATE_FAILED, message);
-    }
-
-    /**
-     * Marks the topic as deleting with the specified actor code and message.
-     *
-     * @param actorCode the actor code indicating the reason for the state
-     * @param message   the message associated with the state
-     */
-    public void markDeleting(LifecycleStatus.ActorCode actorCode, String message) {
-        this.status.update(LifecycleStatus.State.DELETING, message, actorCode);
-    }
-
-    /**
-     * Marks the topic deletion as failed with the specified message.
-     *
-     * @param message the message associated with the failure
-     */
-    public void markDeleteFailed(String message) {
-        this.status.update(LifecycleStatus.State.DELETE_FAILED, message);
-    }
-
-    /**
-     * Marks the topic as inactive.
-     *
-     * @param actorCode the actor code indicating why the topic is being marked as inactive
-     * @param message   the message for the action
-     */
-    public void markInactive(LifecycleStatus.ActorCode actorCode, String message) {
-        this.status.update(LifecycleStatus.State.INACTIVE, message, actorCode);
-    }
-
-    /**
-     * Checks if the topic is active.
-     *
-     * @return true if the topic is in CREATED state, false otherwise
-     */
-    @JsonIgnore
-    public boolean isActive() {
-        return this.status.getState() == LifecycleStatus.State.CREATED;
-    }
-
-    /**
-     * Checks if the topic's current state is retriable.
-     * A state is considered retriable if it is either CREATE_FAILED or DELETE_FAILED.
-     *
-     * @return true if the topic's state is retriable, false otherwise
-     */
-    @JsonIgnore
-    public boolean isRetriable() {
-        return this.status.getState().isRetriable();
     }
 }
